chore: move getDeplooyGatewayOptions to deploy-utils
